프로젝트 2차 리팩토링(전적 카드 로직,구조 변경)
