+++ src/components/SummonerContentContainer.tsx
@@ -1,13 +1,10 @@
-import {
-  getChampionsData,
-  getInitialRecordData,
-  getRunesData,
-  getSummonerData,
-  getVersionsData,
-} from '@/service/requestJsonData.api';
+import { getSummonerData } from '@/service/requestJsonData.api';
 import { GameType } from '@/types';
 import SummonerLeagueContainer from './SummonerLeagueContainer';
-import SummonerRecordClientContainer from './SummonerRecordClientContainer';
+
+import { Suspense } from 'react';
+import SsummonerRecordListContainer from './skeleton/SummonerRecordListContainer.skeleton';
+import SummonerRecordServerContainer from './SummonerRecordServerContainer';
 
 interface SummonerContentContainerProps {
   summonerName: string;
@@ -21,21 +18,10 @@ export default async function SummonerContentContainer({ summonerName, gameType 
     <>
       <SummonerLeagueContainer summonerId={id} />
       <div className="flex-1 min-h-[60rem]">
-        <SummonerRecordServerContainer puuid={puuid} gameType={gameType} />
+        <Suspense fallback={<SsummonerRecordListContainer />}>
+          <SummonerRecordServerContainer puuid={puuid} gameType={gameType} />
+        </Suspense>
       </div>
     </>
-  );
-}
-
-async function SummonerRecordServerContainer({ puuid, gameType }: { puuid: string; gameType: GameType }) {
-  const championsData = await getChampionsData();
-  const [latestVersion] = await getVersionsData();
-  const runesDataArr = await getRunesData();
-  const initialRecordData = await getInitialRecordData(puuid, gameType);
-  return (
-    <SummonerRecordClientContainer
-      data={{ championsData, latestVersion, runesDataArr, initialRecordData }}
-      puuid={puuid}
-    />
   );
 }
