+++ src/components/SummonerRecordClientContainer.tsx
@@ -1,27 +1,43 @@
 'use client';
 
-import { MatchDtoType } from '@/types/response';
-import { ChampionsDataType, RunesDataType } from '@/types/staticData';
+import { MatchDtoType, ParticipantDtoType } from '@/types/response';
+import { ChampionsDataType, ItemsDataType, RunesDataType, SummonerSpellDataType } from '@/types/staticData';
 import ContentBox from './common/ContentBox';
 import Image from 'next/image';
 import useInfinityRecordData from '@/hook/useInfinityRecordData';
-
+import DoughnutChartBox from './DoughnutChartBox';
+import PlayChampionList from './PlayChampionList';
+import totalRecordSummary from '@/utils/totalRecordSummary';
+import filterMatchData from '@/utils/filterMatchData';
+import CardLayer from './recordcard/CardLayer';
+import GameSummary from './recordcard/SummaryContent';
+import MainContent from './recordcard/MainContent';
+import ParticipantsListContent from './recordcard/ParticipantsListContent';
+import Loading from '@/assets/icons/loading.svg';
 interface SummonerRecordClientContainerProps {
   data: {
-    championsData: ChampionsDataType[];
-    latestVersion: string;
-    runesDataArr: RunesDataType[];
-    initialRecordData: MatchDtoType[];
+    champions: ChampionsDataType[];
+    summonerSpells: SummonerSpellDataType[];
+    runesArr: RunesDataType[];
+    items: ItemsDataType;
   };
   puuid: string;
+  initialRecordData: MatchDtoType[];
+  latestVersion: string;
 }
 
-export default function SummonerRecordClientContainer({ data, puuid }: SummonerRecordClientContainerProps) {
-  const { championsData, latestVersion, runesDataArr, initialRecordData } = data;
+export default function SummonerRecordClientContainer({
+  data,
+  puuid,
+  initialRecordData,
+  latestVersion,
+}: SummonerRecordClientContainerProps) {
+  const { champions, runesArr } = data;
   const {
     data: refetchData,
     isFetching,
     refetchNextPage,
+    hasNextPage,
   } = useInfinityRecordData({ initialData: initialRecordData, puuid });
 
   if (!initialRecordData.length) {
@@ -36,15 +52,89 @@ export default function SummonerRecordClientContainer({ data, puuid }: SummonerR
       </ContentBox>
     );
   }
-  console.log(refetchData);
+  const recordSummaryData = totalRecordSummary(refetchData, puuid);
   return (
     <>
-      <button type="button" onClick={() => refetchNextPage()}>
-        다음
-      </button>
-      {refetchData.map((item, index) => (
-        <div key={index}>{index}</div>
-      ))}
+      <ContentBox titleText="최근 게임" css="mb-[2.4rem]">
+        <div className="py-[1.2rem] px-[1.6rem] flex gap-[4.8rem] text-[1.4rem] text-color-gray-500">
+          <DoughnutChartBox win={recordSummaryData.win} lose={recordSummaryData.lose} />
+          <div className="self-center">
+            <div className="text-[1.2rem] text-color-gray-500">
+              <span>{recordSummaryData.avgKill}</span> /{' '}
+              <span className="text-color-red-600">{recordSummaryData.avgDeath}</span> /{' '}
+              <span>{recordSummaryData.avgAssists}</span>
+            </div>
+            <div className="text-[2.4rem] text-color-gray-900">
+              <strong>{recordSummaryData.totalAvg} : 1</strong>
+            </div>
+          </div>
+          <div>
+            <div className="mb-[1.6rem]">많이 플레이한 챔피언 TOP 3 ({refetchData.length}게임)</div>
+            <PlayChampionList
+              filterData={filterMatchData(refetchData, puuid)}
+              championsData={champions}
+              version={latestVersion}
+            />
+          </div>
+        </div>
+      </ContentBox>
+      <div className="flex flex-col gap-[0.8rem] mb-[1.6rem]">
+        {refetchData.map((record, index) => {
+          const summonerGameInfo = record.info.participants.find(
+            participant => participant.puuid === puuid,
+          ) as ParticipantDtoType;
+          return (
+            <div key={index} className="flex flex-col gap-[0.8rem]">
+              <CardLayer
+                isWin={summonerGameInfo.win}
+                gameVersion={record.info.gameVersion}
+                perks={summonerGameInfo.perks}
+                runesDataArr={runesArr}
+              >
+                <GameSummary
+                  isWin={summonerGameInfo.win}
+                  queueId={Number(record.info.queueId)}
+                  timeInfo={{ gameCreation: record.info.gameCreation, gameDuration: record.info.gameDuration }}
+                />
+                <MainContent
+                  participant={summonerGameInfo}
+                  version={latestVersion}
+                  gameVersion={record.info.gameVersion}
+                  data={data}
+                />
+                <ParticipantsListContent
+                  participants={record.info.participants}
+                  puuid={puuid}
+                  championsData={champions}
+                  version={latestVersion}
+                />
+              </CardLayer>
+            </div>
+          );
+        })}
+      </div>
+      {hasNextPage ? (
+        <button
+          className={`h-[5rem] text-[1.6rem] bg-color-gray-00 border border-color-gray-300 w-full rounded-[0.4rem]`}
+          onClick={() => refetchNextPage()}
+          disabled={isFetching}
+        >
+          {isFetching ? (
+            <div className="loader">
+              <Loading />
+            </div>
+          ) : (
+            '더보기'
+          )}
+        </button>
+      ) : (
+        <div className="text-[1.6rem] text-center">
+          마지막 전적입니다. <br />
+          <span className="text-[1.4rem] text-color-gray-400">
+            (YOO.GG는 LOL최신 버전에서 이전5버전까지 제공합니다 )
+          </span>
+        </div>
+      )}
     </>
   );
 }
