+++ src/components/recordcard/RecordDetailContainer.tsx
