+++ src/utils/filterMatchData.ts
@@ -1,5 +1,7 @@
-import { MatchDtoType } from '@/types/response';
+import { MatchDtoType, ParticipantDtoType } from '@/types/response';
 
 export default function filterMatchData(matchData: MatchDtoType[], puuid: string) {
-  return matchData.map(data => data.info.participants.find(participant => participant.puuid === puuid));
+  return matchData.map(data =>
+    data.info.participants.find(participant => participant.puuid === puuid),
+  ) as ParticipantDtoType[];
 }
